+++ Frontend.DB.EF/Models/Discount.cs
@@ -10,7 +10,7 @@ namespace Frontend.DB.EF.Models
         public double Value { get; set; }
         public DateTime? EffectiveDate { get; set; }
         public DateTime? ExpireDate { get; set; }
-        public int UseStatus { get; set; }
+        public byte UseStatus { get; set; }
         public DateTime CreateTime { get; set; }
         public DateTime UpdateTime { get; set; }
